--- src/index.mts
@@ -1,1 +1,0 @@
-export { getWebsiteRank, WebsiteRank } from "./rank.mjs";
