+++ src/internal/arguments.test.ts
@@ -8,7 +8,7 @@ jest.unstable_mockModule("fs", () => ({
 
 describe("parse program arguments and options", () => {
   it("should parse nothing", async () => {
-    const { ArgumentsParser } = await import("./arguments.mjs");
+    const { ArgumentsParser } = await import("./arguments.js");
     const parser = new ArgumentsParser();
     const cmd = "node test github.com";
 
@@ -21,7 +21,7 @@ describe("parse program arguments and options", () => {
   });
 
   it("should parse keywords correctly", async () => {
-    const { ArgumentsParser } = await import("./arguments.mjs");
+    const { ArgumentsParser } = await import("./arguments.js");
     const parser = new ArgumentsParser();
     const cmd = "node test github.com --keywords github gihtub";
 
@@ -34,7 +34,7 @@ describe("parse program arguments and options", () => {
   });
 
   it("should parse file correctly", async () => {
-    const { ArgumentsParser } = await import("./arguments.mjs");
+    const { ArgumentsParser } = await import("./arguments.js");
     const parser = new ArgumentsParser();
     const cmd = "node test github.com --file keywords.txt";
 
@@ -47,7 +47,7 @@ describe("parse program arguments and options", () => {
   });
 
   it("should parse arguments and options correctly", async () => {
-    const { ArgumentsParser } = await import("./arguments.mjs");
+    const { ArgumentsParser } = await import("./arguments.js");
     const parser = new ArgumentsParser();
     const cmd =
       "node test github.com --keywords github gihtub --file keywords.txt --max-page 7";
