--- src/internal/index.mts
@@ -1,2 +1,0 @@
-export { Arguments, ArgumentsParser } from "./arguments.mjs";
-export { formatKeywordRank } from "./format.mjs";
